* Normalize images read from the filesystem so the floats are in the [0,1] range.
I _believe_ this is the expected range for the VGG16 model.

A better solution could be to perform a specific normalization step depending on the initial type and depth of the source image. This change is sufficient for me for now, as it allows the testing of regular JPG files whose ranges were in the [0, 255] range.